requirements.txt
streamlit==1.*
pandas==2.*
altair==5.*
sqlalchemy==2.*
pyathena[sqlalchemy]==3.*
python-dateutil==2.*
boto3>=1.34,<2
